+++ app/Http/Controllers/PackageSizeController.php
@@ -6,9 +6,21 @@ use App\Models\PackageSize;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 
 class PackageSizeController extends Controller
 {
+    public $user;
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::user();
+            return $next($request);
+        });
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,6 +28,20 @@ class PackageSizeController extends Controller
      */
     public function index()
     {
+        if(!$this->user->can('package-size.view')){
+            abort(403, 'sorry! Access Denied');
+        }
+
+        return view('pages.product.package-size');
+    }
+
+
+    public function all_package_size()
+    {
+        if(!$this->user->can('package-size.view')){
+            abort(403, 'sorry! Access Denied');
+        }
+
         return PackageSize::with('unitTypes')->paginate(10);
     }
 
@@ -37,6 +63,10 @@ class PackageSizeController extends Controller
      */
     public function store(Request $request)
     {
+        if(!$this->user->can('package-size.create')){
+            abort(403, 'sorry! Access Denied');
+        }
+
         $id = $request->id;
         $request->validate([
             'unit_type' => 'required',
@@ -94,6 +124,10 @@ class PackageSizeController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!$this->user->can('package-size.edit')){
+            abort(403, 'sorry! Access Denied');
+        }
+
        $request->validate([
                 'unit_type' => 'required',
                 'name'      => 'required|unique:package_sizes,name,' .$id,
@@ -128,6 +162,10 @@ class PackageSizeController extends Controller
      */
     public function destroy($id)
     {
+        if(!$this->user->can('package-size.delete')){
+            abort(403, 'sorry! Access Denied');
+        }
+
        PackageSize:: find($id)->delete();
 
         return response()->json([
